Vigilus/requirements.txt
requests
feedparser
pyyaml
python-dotenv
python-docx